requirements.txt
python-telegram-bot
telegram
numpy
python-dotenv
matplotlib
networkx
requests
bs4
Fraction